fix(ci): split oversized lint targets
